Rename Searchers module file to searchers.rb
